more violet cleanup
